Fcst Component and Landing
Forecast Tile component with size variant, landing page started, footer added
