+++ src/components/blocks/CampusWeatherDetail/CurrentConditions/CurrentValue/CurrentValue.module.scss
@@ -2,7 +2,7 @@
 	flex: 1 1;
 	text-align: center;
 	.label {
-		font-size: 12px;
+		font-size: 14px;
 	}
 	.value {
 		font-size: 18px;
